+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/mediacodec/AsynchronousMediaCodecAdapter.java
@@ -114,7 +114,11 @@ import java.nio.ByteBuffer;
               new AsynchronousMediaCodecBufferEnqueuer(codec, queueingThreadSupplier.get());
         }
         codecAdapter =
-            new AsynchronousMediaCodecAdapter(codec, callbackThreadSupplier.get(), bufferEnqueuer);
+            new AsynchronousMediaCodecAdapter(
+                codec,
+                callbackThreadSupplier.get(),
+                bufferEnqueuer,
+                configuration.loudnessCodecController);
         TraceUtil.endSection();
         if (configuration.surface == null
             && configuration.codecInfo.detachedSurfaceSupported
@@ -157,14 +161,20 @@ import java.nio.ByteBuffer;
   private final MediaCodec codec;
   private final AsynchronousMediaCodecCallback asynchronousMediaCodecCallback;
   private final MediaCodecBufferEnqueuer bufferEnqueuer;
+  @Nullable private final LoudnessCodecController loudnessCodecController;
+
   private boolean codecReleased;
   private @State int state;
 
   private AsynchronousMediaCodecAdapter(
-      MediaCodec codec, HandlerThread callbackThread, MediaCodecBufferEnqueuer bufferEnqueuer) {
+      MediaCodec codec,
+      HandlerThread callbackThread,
+      MediaCodecBufferEnqueuer bufferEnqueuer,
+      @Nullable LoudnessCodecController loudnessCodecController) {
     this.codec = codec;
     this.asynchronousMediaCodecCallback = new AsynchronousMediaCodecCallback(callbackThread);
     this.bufferEnqueuer = bufferEnqueuer;
+    this.loudnessCodecController = loudnessCodecController;
     this.state = STATE_CREATED;
   }
 
@@ -181,6 +191,9 @@ import java.nio.ByteBuffer;
     TraceUtil.beginSection("startCodec");
     codec.start();
     TraceUtil.endSection();
+    if (Util.SDK_INT >= 35 && loudnessCodecController != null) {
+      loudnessCodecController.addMediaCodec(codec);
+    }
     state = STATE_INITIALIZED;
   }
 
@@ -273,6 +286,9 @@ import java.nio.ByteBuffer;
             codec.stop();
           }
         } finally {
+          if (Util.SDK_INT >= 35 && loudnessCodecController != null) {
+            loudnessCodecController.removeMediaCodec(codec);
+          }
           codec.release();
           codecReleased = true;
         }